requirements.txt
streamlit
openai
python-docx
python-dotenv
PyPDF2
requests
beautifulsoup4